Adds smart partition reassignment
When persistence is enabled and a crashed
member is rejoining, a stored partition
table snapshot may be picked during
repartitioning task, with a view to minimize
partition rearrangement and migrations.
An eligible partition table snapshot must
include replica assignments to current members
and not any of the shutting down members.
If multiple such candidates are located,
then the one with least "distance" from
the current partition table is chosen.
Distance calculation is the sum of all
partitions' distance. For a given partition
assignment, the distance calculation considers
replica index assignments. If a given
replica index is assigned to the same member
in both partition assignments, then their
distance is 0. If a member is assigned to
replica index 0 and index 2 in the compared
partitions, then the distance is incremented by 2.
A replica index that is assigned in one
but unassigned (null) in the other partition
increments distance by MAX_REPLICA_COUNT.