Add compatibility tests for the enums used in the protocol
We were not verifying the compatibility of the enums (or enum-like constants)
used in the protocol. This might cause us to miss the extensions to
enums used in the protocol, as the changes might cause compatibility
problems for our clients.

These tests use hardcoded values for the enums used in the protocol
as well as hardcoding the expected enum members so that we can
catch modifications to values or addition/deletion of enum members.

There was no single type used in the protocol definitions for enums
so I had to go over the all protocol definition parameters that
have the type integer, byte, short, or string which I believe
are the only types we use to represent enums.

I also tried to convert the enum-like constants that we use in protocol
to an enum wherever possible to have a better capability to
check possible modifications in the future.