Isolate WAN clusters in tests (#19216)
WAN tests use the mock network that has `MockJoiner` joiner
implementation that tries to join all members created with the same
factory. This change extends the functionality of `MockJoiner` with
support for cluster isolation, specifically for WAN tests. If isolation
is enabled by a test - disabled by default, no tests should be affected
by this change apart from WAN tests - `MockJoiner` prevents members
outside of a port range from attempting to join to each other. With
other words, this change makes `MockJoiner` to behave like `TcpJoiner`
that looks up for possible cluster members within a limited port range
only.

Attempting to join members of WAN source and target clusters may fail
WAN sync and consistency check tests in edge cases. If there is an
ongoing join request - that establishes a connection - when the sync or
consistency check request is made, the request will find a connection
between the parties and will try to perform the requested action over
that connection. But the connection will be closed once the two members
realize they are members of two separate clusters. This closes the
connection under the ongoing WAN sync/consistency check operation and
fails the test.