Fix JetClassLoader not set as context classloader (#19264)
- when we introduced the processor classloader feature we introduced a
regression that set the context classloader to null, this is #19083

- discovered cases, where the job classloader wasn't set as the context
classloader for some methods - most importantly Processor#close, this
could cause issues when e.g. closing some resources needs to load a new
class and it uses the context classloader for that

- setting the context classloader for some methods, e.g.
Processor#isCooperative is likely unnecessary, but it is easy enough to
do to keep the consistency

Fixes #19083

* Store instance of JetServiceBackend in ExecutionContext

The NodeEngineImpl is used to get the JetServiceBackend service, it is
not possible to get the service from it in completeExecution when the
HazelcastInstance instance is being shut down.