Fix composite index tests and improve query engine error reporting
OpenJ9 is apparently less aggressive with optimizations and that's why
the test was only (mostly?) failing in OpenJ9 builds. The failure is
specific to the test itself: the custom predicate used in the test
didn't expect it can be invoked concurrently across threads.
As a result, `ClassCastException` was produced, which then was silenced
in `QueryEngineImpl`.

Also, this PR improves error reporting in `QueryEngineImpl` when
`disableMigrationFallback` is `true` (used for testing purposes only).
Now a proper exception is reported to a caller instead of a cryptic
"could not execute query for all partitions" exception.

Cherry-picked from e7ca38d41c21917bdd86b4cf912f284c097187a1

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/4080
Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/4091
Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/4193