Cache byte array used to calculate partition stamps (#19311)
During repartitioning, the PartitionStamp#calculateStamp method may be
called many times and each invocation creates a byte array buffer with
enough size to contain partitionCount integers.

By running `WanCounterMigrationTest#testCountersReachZeroAfterMigrateToNewAndBack`
and setting the partition count to 2053, this is how many times it was
called from three different methods:
- com.hazelcast.internal.partition.impl.PartitionStateManager.updateStamp
16428
- com.hazelcast.internal.partition.impl.InternalPartitionServiceImpl.createPartitionStateInternal
13
- com.hazelcast.internal.partition.PartitionTableView.stamp
16

As Zoltan says- if my calculation is right, with ~7k partitions a
migration allocated ~436MB only in this method with 2053 partitions it's
 ~38MB

That places considerable GC pressure. We will then create a dedicated
buffer in PartitionStateManager and reuse it every time we need to
calculate the stamp, this way avoiding most of the allocations.

We also use a IntHashSet in MigrationManager to consume less space.