Add support for declarative compact serialization configuration (#19276)
* Add support for declarative compact serialization configuration

This PR adds support for declarative configuration for the compact
serialization format.

For the XML configuration, the structure will look like

```xml
<serialization>
	<compact-serialization enabled="true">
		<registered-classes>
			<class>com.foo.Bar</class>
			<class type-name="baz" serializer="com.foo.BazSerializer">com.foo.Baz</class>
		</registered-classes>
	</compact-serialization>
</serialization>
```

For YAML, the structure is as follows

```yaml
serialization:
	compact-serialization:
		enabled: true
		registered-classes:
			- class: com.foo.Bar
			- class: com.foo.Baz
			  type-name: baz
			  serializer: com.foo.BazSerializer
```

After an internal discussion, we decided to remove support for
registering reflective serialization with a custom type name and
registering explicit serializer without a type name, and that
changes are reflected in the declarative configuration as well.

Since this is a BETA feature, I didn't make any changes to the
XSDs. That's why one has to disable schema validation to be
able to use declarative configuration. This limitation will be
documented in the reference manual.

* set flag to correct value in ExternalizableEmployeeDTO

* - add compact serialization config to ConfigCompatibilityChecker
- add a TODO to the test cases
- remove the unnecessary newline