Fix processor classloader related issues (#19182)
* Fix processor classloader issues on windows

The ChildFirstClassLoader is a subclass of URLClassLoader,
which needs to be closed. This was the case in
ChildFirstClassLoaderTest.

Also added call to close to com.hazelcast.jet.impl.JobExecutionService.completeExecution
when the processor classloaders are removed.

Fixes #19173, #19174

* Remove classloaders after closing PMS on coordinator node

* Add tests for job CL for ProcessorSupplier and PMS

There were no tests to check if job classloader is set for init, get and
close methods of ProcessorSupplier and ProcessorMetaSupplier.

This commit adds the tests and fixes places where the classloader was
set incorrectly.