Collect incoming queues metrics before draining queue into Inbox
Justification:
We drain into Inbox only when the Inbox is empty. So each
invocation of `fillInbox()` drains a lot of entries from
the incoming queues. If we collect queues metrics immediately
AFTER draining them then the queues will always appear empty,
even for busy processors -> We need to collect entries BEFORE
draining the queues.

I'm not sure how to test this, ideas are welcome.