Handle undefined cleanup order for PS and PMS
The execution completion is done asynchronously so it might happen that
MasterJobContext#finalizeJob, which closes PMS and removes the job class
loader, is done before JobExecutionService#completeExecution, which then
tries to access already removed classloader.

This introduces a reference counter when removing job classloaders.

Also refactored the classloader handling to a separate service.

Fixes #19381
Fixes #19001 (the second occurrences related to this)