Consolidate UUID in MC WAN events (#19422)
Set UUID consisteltly on all WAN MC events. The change sets the UUID in the WAN anti-entropy ignored events. Also, at some places we did not emit the ignored events: if there was no map found to be synchronized or the map was not configured for WAN replication, the ignored events were not sent to MC.

Additionally, the configuration related events are extended with a UUID field to be consistent with the other WAN events. In these events the UUID doesn't correlate with any field on the response, MC needs it only for grouping.

Implements hazelcast/hazelcast-enterprise#4154