Use higher number ports in TestManagedContext (#19447)
Because the tests of different modules run simultaneously with each
other, sometimes these managedContext tests cannot bind to the ports
5701 and 5702, and when that happens, the tests fail since cluster
is not formed. When we don't explicitly specify the port numbers,
they try multiple number of ports and find each other. Since the
cluster name is unique for this test, it does not interfere with
other tests exactly.