Test portable array mapped as OBJECT in SQL (#19451)
The portable resolver currently checks that the resolved and assigned
type are same. For arrays we resolve OBJECT, so we check that the user
actually used OBJECT.

However, in a future release we plan to introduce support for arrays and
for nested objects. We also want to preserve backwards compatibility.
Therefore this PR adds a test that checks that mapping ARRAY as OBJECT
still works.