Warn if job jar contains core hazelcast classes (#19426)
Each jar submitted with `hz-cli` is scanned for the presence of a specific Hazelcast class (currently `HazelcastBootstrap`). When it's found a warning will be printed.

The jar is scanned against a class name regardless of package name to detect shaded Hazelcast dependency as well.

Fixes https://github.com/hazelcast/hazelcast/issues/19426