Do not check for tombstones in ReplicatedMap (#19005) (#19519)
Reasoning:
Tombstones are a left-over from the old (before-2015) implementation of ReplicatedMap.
The old implementation supported updates on an arbitrary member and used VectorClock
to detect conflicts. Tombstones were needed because a plain entry removal would lose
the VectorClock state.

The current implementation does NOT use VectorClock at all. It routes all updates via
a partition owner which acts as a serialization point.

This is the PR introducing the current implementation: https://github.com/hazelcast/hazelcast/pull/6176

Given tombstones are never used we can also simplify the `repmap.size()` method - it no longer
has to check each record for being a tombstone or not. This improves time complexity of the `size()`
method from `O(n)` to `O(1)` where `n` is a size of a replicated map.

I removed the tests, which created replicated map records with a null
value. The old implemenation used this to indicate a record is a
tombstone, but this state is not possible in the current implementation.
Other tests are passing.

(cherry picked from commit 5cfce66e216d2d2a486e34cea35c4f6a8f48e51a)