Don't run onClusterRestart on failover code
Added explanation for non-trivial methods of the ClientClusterService.

renamed onClusterChange to beforeClusterChange and
renamed onClusterRestart to afterClusterRestart
to make it clear when they are called. Also added javadoc for those.

While writing javadoc, it become more clear that onClusterRestart
should not be called on failover client. Refactored that as well.