Throw unsupported operation exception if perEntryStatsEnabled is false (#19483)
Issue:
Issue happens if you configure both ttl and maxIdle for a key. In this case, we need to know last-
update-time to calculate ttl expiry time which is needed to compare with maxIdle expiry time.

Modifications:
For 4.2, this requirement can only be satisfied when perEntryStatsEnabled is enabled. So we will 
throw exception and warn user to set perEntryStatsEnabled to true.

Note:
We cannot add last-update-time to expiry metadata due to the backward compatibility issues.