FieldKind exposed as public API [API-437] (#19517)
* FieldID exposed as public API

Introduced FieldID instead of FieldType which will be a
more correct API to show field id's of Compact and Portable.

FieldType has some unfortunate methods where typeSize needs to be
asnwerd correctly. Since they differ between Portable and Compact,
we can not go with the same API.

The plan is to deprecate/remove FieldType together with Portable.