Add Nullable Primitives to Compact
New Nullable* methods are added to GenericRecord,
CompactReader,CompactWriter.

Portable throws UnsuportedOperationException for new API's on
GenericRecord.

SQL side has changed to use NULLABLE versions to support inserting
null in SQL statements. Note that the main reason of this addition to
Compact is SQL.

Time/Date/Timestamp and Timestamp With Timezone is Nullable by default. We have removed the fixed size versions of those for now.

readObject/Array methods are renamed as readCompact/Array.

The methods with defaultValue are removed from CompactReader for Nullable kinds. They are left for only nonnull primitive kinds.