Fix Ringbuffer test failures
Fixes #19696
ReliableTopicDestroyTest.whenDestroyedThenRingbufferRemoved
Created a simpler reproducer RingbufferDestroyTest.whenDestroyAfterAdd_thenRingbufferRemoved
The cause was recreation of the RingbufferContainer in ReadOneOperation.getWaitKey
This also addresses review comment from #19630.

Fixes #16469
RingbufferAddAllReadManyStressTest.whenShortTTLAndBigBuffer
The stress test is incorrect, the ReadManyOperation doesn't throw the
StaleSequenceException when head is stale and the items are just missing
in the result. This was introduced in #16303.

Also fixed HashMap->ConcurrentHashMap in RingbufferService.
This Map is modified from operation thread when RingbufferContainer is
created and also from destroyContainer, which may run directly on the
user's thread when the ringbuffer is local on the member.