Complete job execution only once
There are calls to `completeExecution` from multiple places.
When there is a failure both on coordinator and other member,
but the coordinator is slow to process its own failure it's
possible that we end up with the following log

```
... calling completeExecution because execution terminated before it started
```

and call `completeExecution` twice.

It's the only place where we remove the ExecutionContext from `executionContexts`,
so we can use it to check if it runs for 2nd time.

Fixes #19262