Fix protocol verification bug (#19778)
Tests were failing because SingleProtocolDecoder#verifyProtocol is
called with not only protocol header but also subsequent data in
the case that verifyProtocol fails at first.
Instead of running the verifyProtocol only once with the initial
protocol data, we run it again and again, and when we encounter a byte
sequence that matches the protocol header by chance, this was
accidentally causing protocol to be verified with a data that is not in
the protocol header before the connection closed in the encoder side.
It became possible after we have deferred the mechanism of connection
closing exception throwing to the encoder side.

Fixed it by adding a boolean flag to ensure that this verifyProtocol
method will only be called with the first incoming protocol header

--- Also
* Pretend to process source buffer in SingleProtocolDecoder in case
of that protocol verification fails

Otherwise, TLSDecoder wasn't get out of its dirty loop although the
channel was closed. This was happening in the not verified protocol
case.

Co-authored-by: Ramiz Dündar <40149584+ramizdundar@users.noreply.github.com>
(cherry picked from commit 81956a7f970f2b2b0aaa9f075007d08b36a00ff3)