Wait for safe state in SplitBrainTest [5.0.z]
In #19562 we started new instance at the end of the test so the job
might restart and be stopped correctly.

Wed need to wait for migrations to finish before we exit the test,
otherwise the test cleanup will try to ditch jobs before they are
restarted (currently it's impossible to ditch job which is restarting).

Fixes #19560 in 5.0.z