Return values in the unsigned byte range while reading from positions in `UnsafeObjectDataInput`
The contract of the `InputStream#read` says that the return values
must be in the unsigned byte range (0 to 255). We do that in our
`read()` methods, but, the conversion to the required range was missing
in `read(int)` in `UnsafeObjectDataInput`.

Note that, `read(int)` is defined in our own interface `BufferObjectDataInput`
which does not have any documentation describing its contract, but
it must be consistent with `read`.