Remove support for char and char[] from Compact
We have decided to not support `char` and `char[]` types on the compact
given that it is not a common type in other languages and we plan
to add support for unsigned integers.

It is true that for the Java users this might be a drawback but we think
`char` type is not widely used and currently, it is not possible to
use it in other parts of the Hazelcast, such as SQL. Hence, we think
our decision to remove it is justified.