Fix QueueStatisticsTest#testAge
Looking at the logs of the test failures, it seems max age, min age,
and average age were equal to their default values (0, Long.MAX_VALUE, 0)
despite there being offer and poll operations.

That could only happen if in the `QueueContainer#age` method, we calculate
the elapsed time as some non-positive value.

So, this attempts to fix this problem by sleeping for a while between offer
and poll operations.

Note: I was thinking of changing the if condition on the aforementioned from
`elapsed <= 0` to `elapsed < 0`. If you think it makes sense, I could also
do that.