Reduce re-creation of ExecutionContext after job failure
We used to re-create the ExecutionContext, if a data packet is received
after the execution terminated. This is quite likely to happen after a
failure or cancellation as executions are concurrently completing on
members (in 10% of cases in local tests).

We greatly mitigate by storing the failed execution IDs for 5 seconds
after they fail. The timeout is short reduce the amount of work in
query-heave scenarios, if it's not enough the old mechanism still
applies.

Fixes #19897