Fix possibly infinite loop in handling index splits (#19966)
SQL provides migration tolerance for index scans. When migration
happens, the same processor instance will handle the partition after
it's moved, making a remote operations to the new owner. Internally
it's implemented by using splits. Initially the processor starts with
one split, and if migrated partition is detected, it splits the initial
"split" into multiple ones, based on the number of new owners we now
need to talk to read data for the assigned partitions.

In the split-brain situation it happened that the Index-read operation
threw MissingPartitionException, however the PartitionService still
reported the old owner. Therefore the `splitOnMigration()` method didn't
actually split the partitions, but the same single split was used. And the
algorithm tried the new split again. And this went on infinitely.

The fix is that we don't try the split again, but continue with the next
split. This breaks the loop and the job can be cancelled. The old
implementation broke the requirement that the processor must return
control from time to time.

The problem reproduced only on 5.0, not on 5.1. I didn't investigate
why, but probably the partition owner is changed earlier and the
infinite loop terminated. We'll forward-port the fix to 5.1 anyway to
make the code more robust.

Fixes #19472