Fix the final position of a fixed size compact object
While reading a compact object from a stream of buffers
the end position was not set correctly if the compact is
consisting of only fixed size objects.
In this pr, we make sure that the position is moved to and of the
compact object so that next item can be read correctly.

related to https://github.com/hazelcast/hazelcast/issues/19952