Remove Closable from BufferObjectDataInput (#19977)
* Remove Closable from BufferObjectDataInput

We have a new implementation of serialization which keeps the
reference of BufferObjectDataInput to read it from lazily instead
of reading it on deserialization.

The problem is that we are closing the BufferObjectDataInput in
several places where the lazy implementations can not work anymore.

As a solution, I removed Closeable from BufferObjectDataInput
after we realize that all the implementations we have just sets
internal buffers to null. There are no resources cleaned up,
therefore removing `close` implementations seems safe.

On other important change in the pr is in the `TaskletExecutionService`.
In `SqlClientCompactQueryTest` we are setting a `FilteringClassLoader`
to the hazelcast config. This classLoader is used by all the threads
created by hazelcast except cooperative threads of jet.
We have realized it when we can not reproduce the issue with single jvm.
The reason is turned out to be config class loader is not used by the
jet cooperative threads. As a solution:
1. Config class loader is set to context class loader of cooperative
threads.
2. Overriding the context class loader is prevented if jobClassLoader
is null. So we keep the behavior same if a class loader is set to
JobConfig. Otherwise, config class loader is used.

fixes https://github.com/hazelcast/hazelcast/issues/19952


Co-authored-by: Viliam Durina <viliam@hazelcast.com>