Better toString for arrays in peek (#19985)
The default toString method for arrays returns something like
`[Ljava.lang.String;@f6f4d33`. This commit uses `Arrays.toString()`
instead.

This is not 100% b-w compatible, the logged text can potentially be very
large, and it's not the same as in previous version, but we still do it
for these reasons:

- the exact output format wasn't specified

- peek functions are meant for development. We don't expect them to be
used for production (though it's possible)

- the old format wasn't very useful for automatic parsing, maybe only
for checking instance equality.