Remove Lazy GenericRecord implementations
The lazy implementation DefaultCompactReader was not
threadsafe. The underlying position is modified at each
read attempt where concurrent reads can not work.

An attempt to solve this problem is here
https://github.com/hazelcast/hazelcast/pull/20040
See commit messages of that to see details.
In short, the code becomes unnecessarily complex.

In this pr, the behaviour is changed so that we
return a completely deserialized GenericRecord
to user via API's like map.get()/EntryProcessor.

The query path still uses the lazy implementation. We do
not deserialize whole data still.

When in memory format is Object and we don't have the classes
on the server. The stored objects where lazy generic records.
And the query over them where working smoothly because the
GenericQueryReader works with InternalGenericRecord already.
With this change, the stored objects are no longer InternalGenericRecord.
They are just GenericRecords. Solution is to make
DeserializedGenericRecord an instance of InternalGenericRecord.