Fix SplitBrainTest#when_newMemberIsAddedAfterClusterSizeFallsBelowQuorumSize_then_jobRestartDoesNotSucceed
The source of the failure is assertTrueEventually. Because it's only
sufficient to assert conditions that eventually become true and then
always remain true. It tries to run assert task 5 times per second, and
can miss if a condition is true at one period time and then false again
(anyway, there is no chance to try it continuously). See:
https://github.com/hazelcast/hazelcast/blob/master/hazelcast/src/test/java/com/hazelcast/test/HazelcastTestSupport.java#L1253
In this test, although job starts to run (job status becomes RUNNING
for some period of time), it completes very quickly, so assert tasks
trials may not coincide with this RUNNING state.