Fix SuspendExecutionOnFailureTest assertions
I think the assertion on line 192 was wrong. If a vertex fails after a
certain item, there's no guarantee that a downstream vertex already
processed all previous items. The test failed rarely because the source
emits items one by one with 5ms delay, so it was very likely that all
items were sunk, but if there was some hiccup, it could fail.

Fixes #19836