Simplify ExecutionEvalContext
We no longer need the interface. The IMDG implementation with more
dependencies was already removed and the Jet's impl was simple. We also
don't need the test implementation because it was the same as the Jet's
impl, except that it used test serialization service - the test
implementation was replaced with a simple static method in
SqlTestSupport.