Make Processor.Context available from ThreadLocal (#20311)
In the SQL module we need to access Processor context from the various
lambdas. It's a PITA to add it where supported (we have to use factories
from which we get the actual function), such as in `mapUsingContext`.
But in many places it's not supported, e.g. in `AggregateOperation`
methods.

This PR replaces the metrics context that already was in a thread local
with processor context.