[HZ-760] Configuration update & reload (#20260)
Configuration update is updating the runtime configuration of each
member in the cluster with a new configuration given. This action may
include configuration rewriting at the end.

Configuration reload is configuration updating, in which the given
configuration is read from the declarative configuration.

This is an external API. It is reachable from REST API. Also in the
future it could be accessed from MC. There are two types of update.
First one is called `update` endpoint and the other is called `reload`
endpoint. These endpoints fall under `RestEndpointGroup.CLUSTER_WRITE`
group and are:

 - /hazelcast/rest/config/reload
 - /hazelcast/rest/config/update

Response to this call is a JSON contains the following: status, message
and result. Status and message is straightforward. Result is however,
another JSON object. It contains added and ignored configs and both are
set of `ConfigNamespace` as JSON. Or it's basically `ConfigUpdateResult`
in JSON format.

Whether they take the new configuration from REST request parameter or
re-reading the declarative configuration, once new configuration is
obtained, these two endpoints have the same logic handled by the
`EnterpriseClusterWideConfigurationService#ConfigUpdater` class. It has
2 phases, `scan` and `apply`. During `scan` it iterates through
dynamically changeable configs and during `apply` it adds the ones which
aren't found in the runtime config. Since `apply` uses existing dynamic
change programmatic API, it inherits all guarantees it provides.
`ConfigUpdater` returns a `ConfigUpdateResult` which includes 2 sets of
`ConfigNamespace`, one for added configs and one for ignored configs.

Please note that for `upload` endpoint, the given config as parameter
doesn't have to be the superset of the runtime configuration.