Allow member rejoin when first arrangement not done
The previous condition to determine if a member is
rejoining the cluster (having a persistent identity)
would only consider if a partition assignments snapshot
exists for that member UUID (meaning that the member
previously left the cluster unexpectedly). If partition
arrangement was not yet done at the time the member
left, there is no snapshot, so it is not recognized as
a rejoining member. If, however, at the time of
rejoining partition arrangement is not yet done, then it
should be admitted as a rejoining member, without having
to perform a force start.